requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.2
boto3==1.39.11
botocore==1.39.11
botocore-stubs==1.40.33
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.116.2
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.8.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.20.3
psutil==7.1.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3fs==2025.9.0
s3transfer==0.13.1
safetensors==0.6.2
scipy==1.16.2
scmrepo==3.5.2
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
timm==1.0.20
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer==0.19.1
types-aiobotocore==2.24.2
types-aiobotocore-cloudformation==2.24.2
types-aiobotocore-dynamodb==2.24.2
types-aiobotocore-ec2==2.24.2
types-aiobotocore-lambda==2.24.2
types-aiobotocore-lite==2.24.2
types-aiobotocore-rds==2.24.2
types-aiobotocore-s3==2.24.2
types-aiobotocore-sqs==2.24.2
types-awscrt==0.27.6
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==4.0
